+++ VentControl2/src/Motor.h
@@ -1,0 +1,42 @@
+#ifndef MOTOR_H
+#define MOTOR_H
+
+#include <Arduino.h>
+#include "nextionDisp.h"
+
+/*  
+Speed value 0-5.
+Calls sysValUpdate() to update display
+It increases or decreases the motor speed gradually until it reaches the value.
+*/
+
+class Motor{
+    private:
+		int speed;
+		int PWM;
+		int dly = 30;            // How fast the motor goes from one value to another
+		int output_pin = 9;
+		int max = 0;
+
+
+        // PWM SETTINGS :
+        static const int s1 = 70;				// motor PWM value at setting 1
+        static const int s2 = 100;				// motor PWM value at setting 2
+        static const int s3 = 150;				// motor PWM value at setting 3
+        static const int s4 = 210;				// motor PWM value at setting 4
+        static const int s5 = 255;				// motor PWM value at setting 5
+
+    public:
+        Motor(int pin);
+        void setSpeed(int); // 
+        int getSpeed();        // Returns the current motor speed (int)
+        void setPin(int);   
+        int getMax();       // @TODO: Not implemented
+
+        bool isRunning = 0;
+
+    private:
+
+};
+
+#endif
