+++ VentControl2/src/filter.h
@@ -1,3 +1,8 @@
+#ifndef FILTER_H
+#define FILTER_H
+
 
 
 float filter(float i, float n, float d, float a);
+
+#endif
