+++ VentControl2/src/heating.cpp
@@ -1,16 +1,4 @@
-#include "motor_speed.h"
-#include "sensor.h"
-
-extern bool enableHeating;
-extern int tempUpper;
-extern int tempLower;
-extern int autoCyckle;
-extern int n;
-extern int k;
-extern Motor motor1;
-extern DigitalTemp t_Outside;
-extern AnalogTemp t_Panel;
-extern AnalogTemp t_HeatedAir;
+#include "heating.h"
 
 void heating(){
 	// Check if heating is enabled
